+++ src/requirements-mac.txt
@@ -2,5 +2,5 @@ coloredlogs==15.0.1
 docker==6.1.3
 mypy_extensions==1.0.0
 PyYAML==6.0.1
-requests==2.31.0
+requests<=2.29.0  # https://github.com/docker/docker-py/issues/3113
 verboselogs==1.7
